+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/template
 
+## 0.2.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.16.6
+
 ## 0.2.6
 
 ### Patch Changes